# ggml: Add POOL2D OP for GPU acceleration to the Vulkan backend in the MobileVLM model.